wishlist

+++ DTOs/OrderItem/OrderItemDto.cs
@@ -5,7 +5,7 @@
 		public Guid ProductId { get; set; }
 		public string ProductName { get; set; } = string.Empty;
 		public string ProductImage { get; set; } = string.Empty;
-		public float ProductPrice { get; set; }
+		public decimal ProductPrice { get; set; }
 		public int Number { get; set; }
 	}
 }
